backend/requirements.txt
# ==========================================================
# CORE DJANGO
# ==========================================================
Django==4.2.11
djangorestframework==3.14.0
django-filter==24.1
django-cors-headers==4.3.1

# ==========================================================
# AUTH & SECURITY
# ==========================================================
djangorestframework-simplejwt==5.3.1
django-cryptography==1.1
cryptography==42.0.5
argon2-cffi==23.1.0
PyJWT==2.8.0

# ==========================================================
# DATABASE
# ==========================================================
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ==========================================================
# CACHE & CELERY
# ==========================================================
redis==5.0.3
celery==5.3.6
django-redis==5.4.0
flower==2.0.1

# ==========================================================
# BLOCKCHAIN / WEB3
# ==========================================================
web3==6.15.1
eth-account==0.10.0
eth-utils==2.3.1
hexbytes==0.3.1

# ==========================================================
# IPFS STORAGE
# ==========================================================
ipfshttpclient==0.8.0a2

# ==========================================================
# ASYNC / WORKERS
# ==========================================================
asgiref==3.7.2
uvicorn==0.29.0
gunicorn==21.2.0

# ==========================================================
# AI ENGINE (Risk / AML)
# ==========================================================
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2

# ==========================================================
# QR CODE PAYMENTS
# ==========================================================
qrcode==7.4.2
Pillow==10.2.0

# ==========================================================
# API & VALIDATION
# ==========================================================
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.3

# ==========================================================
# MONITORING & LOGGING
# ==========================================================
sentry-sdk==1.40.5
structlog==24.1.0

# ==========================================================
# RATE LIMITING / SECURITY HARDENING
# ==========================================================
django-ratelimit==4.1.0
django-axes==6.3.0

# ==========================================================
# UTILITIES
# ==========================================================
python-decouple==3.8
uuid==1.30

# ==========================================================
# TESTING
# ==========================================================
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0